refactor(BA-7563): split full user creation into dedicated user ops
Move FullUserCreation and create_full_user off the shared RBAC write ops onto
UserWriteOps/UserOpsProvider under repositories/ops/user/, injected into the auth
and user db sources. The primitive is broken into the steps it already ran —
user scope, default keypair, domain enrollment, project enrollment. No behavior
change.
